[mod/#89] storeId 인자로 넘기기

+++ feature/store/src/main/java/com/napzak/market/store/store/navigation/StoreNavigation.kt
@@ -18,7 +18,7 @@ fun NavGraphBuilder.storeGraph(
     navigateToUp: () -> Unit,
     navigateToProfileEdit: () -> Unit,
     navigateToProductDetail: (Long) -> Unit,
-    navigateToStoreReport: () -> Unit,
+    navigateToStoreReport: (Long) -> Unit,
     modifier: Modifier = Modifier,
 ) {
     composable<Store> {
